[API-25] Refactor on classes location
